[fix]: race condition on multi-lookup then multi-forward-passes on the same request (#1268)
* dont pin on second lookup


* remove layerwise race condition with lookup unpin


* lookup unpin happening for consumers as well


* make unique lookup ids


* make unique lookup ids


* Union[str, uuid.UUID] for lookup_id


* tested regular inference and move and compress controller endpoints


* change to str(uuid.uuid4()) and all 3 experiments still work


* change optional types in clients


* add on global scoping for unit tests


* small integration test fix


---------
