+++ lmcache/v1/lookup_client/abstract_client.py
@@ -15,14 +15,14 @@ class LookupClientInterface(metaclass=abc.ABCMeta):
     """Abstract interface for lookup clients."""
 
     @abc.abstractmethod
-    def lookup(self, token_ids: torch.Tensor, request_id: Optional[str] = None) -> int:
+    def lookup(self, token_ids: torch.Tensor, lookup_id: Optional[str] = None) -> int:
         """
         Perform lookup for the given token IDs.
 
         Args:
             token_ids: The token IDs to lookup
 
-            request_id: The request ID to associate with the lookup
+            lookup_id: The lookup ID to associate with the lookup
 
         Returns:
             The number of tokens that can be loaded from cache
